fix: comprehensive build fixes for Geode mod
CMakeLists.txt:
- Switch from include(CPM.cmake)/include(Geode.cmake) to add_subdirectory
  pattern matching current Geode SDK and official example-mod
- Add macOS architecture and visibility preset settings
- Use CONFIGURE_DEPENDS in GLOB_RECURSE

mod.json:
- Fix version format: 'v1.0.0' -> '1.0.0' (Geode expects semver without prefix)

MenuLayerHook.cpp:
- Use named $modify(MyMenuLayer, MenuLayer) pattern so menu_selector works
  with custom methods (matches official Geode example-mod)
- Remove m_pfnSelector hack and onMoreGames intercept pattern

PlayLayerHook.cpp:
- Remove unused #include <Geode/modify/PlayerObject.hpp>
- Use named $modify(MyPlayLayer, PlayLayer) for consistency

PathfinderPopup.cpp:
- Replace ButtonSprite with IconButtonSprite (confirmed Geode v4 API)
- IconButtonSprite::create(bg, icon, text, font) signature verified from SDK source
- Remove invalid includes that were cleaned up in previous commit

EditorPauseLayerHook.cpp:
- Use named $modify pattern for consistency
- Remove unnecessary Pathfinder.hpp include

Pathfinder.cpp:
- Remove redundant #include <Geode/loader/Mod.hpp> (already in Geode.hpp)
- Fix implicit conversion warning in std::min<int>

+++ src/ai/Pathfinder.cpp
@@ -1,7 +1,6 @@
 #include "Pathfinder.hpp"
 
 #include <Geode/Geode.hpp>
-#include <Geode/loader/Mod.hpp>
 
 #include <algorithm>
 #include <cmath>
@@ -130,7 +129,7 @@ Action Pathfinder::chooseAction(PlayerState const& s,
             auto nextState = PhysicsModel::step(node.state, act, 0.0f, 480.0f);
 
             // Rough collision check against the obstacle silhouette ahead.
-            int slot = std::min<int>(node.depth, heightsAhead.size() - 1);
+            int slot = std::min<int>(node.depth, static_cast<int>(heightsAhead.size()) - 1);
             float ground = (slot >= 0) ? heightsAhead[slot] : 0.0f;
             float addCost = (nextState.y < ground) ? kCollisionPenalty : 0.0f;
